feat(Route router) - moved overview.js to routes folder - created Collectors.js - created Bins.js - created Maps.js - created Analytics.js (all in routes folder)
References Design Homepage Component #14
