+++ src/App.js
@@ -3,7 +3,12 @@ import React from "react";
 import { Navbar } from "./components/Navbar";
 import { Sidebar } from "./components/Sidebar";
 // import { HumidityDisplay } from "./components/HumidityDisplay";
-import { Overview } from "./components/Overview";
+import { Overview } from "./routes/Overview";
+import { RouterProvider, createHashRouter } from "react-router-dom";
+
+const router = createHashRouter([
+
+]);
 
 export function App() {
   return (
