+++ src/routes/Analytics.js
@@ -1,0 +1,9 @@
+function Analytics() {
+    return (
+        <div>
+            Statistics
+        </div>
+    )
+}
+
+export default Analytics
