+++ src/routes/Bins.js
@@ -1,0 +1,9 @@
+function Bins() {
+    return (
+        <div>
+            List of Bins
+        </div>
+    )
+}
+
+export default Bins
