+++ src/routes/Collectors.js
@@ -1,0 +1,9 @@
+function Collectors() {
+    return (
+        <div>
+            List of Collectors
+        </div>
+    )
+}
+
+export default Collectors
