+++ src/routes/Map.js
@@ -1,0 +1,9 @@
+function Map() {
+    return (
+        <div>
+            This is a beautiful map
+        </div>
+    )
+}
+
+export default Map
