+++ src/styles/Sidebar.css
@@ -135,4 +135,5 @@
   .nav-element svg {
     font-size: 40px;
   }
+ 
 }
